+++ cse3902/Commands/CommandList.cs
@@ -35,9 +35,8 @@ namespace cse3902.Commands
                 {new Keys[] {Keys.Q}, new QuitCommand(game)},
                 {new Keys[] {Keys.R}, new ResetCommand(game)},
                 {new Keys[] {Keys.Up, Keys.Left, Keys.Down, Keys.Right, Keys.W, Keys.A, Keys.S, Keys.D}, new LinkMovementCommand(game) },
-                {new Keys[] {Keys.D1, Keys.D2, Keys.D3, Keys.D4}, new LinkUseItemCommand(game) },
+                {new Keys[] {Keys.D1, Keys.D2}, new LinkAttackItemCommand(game) },
                 {new Keys[] {Keys.D5, Keys.D6, Keys.D7, Keys.D8}, new LinkChangeWeaponCommand(game) },
-                {new Keys[] {Keys.Z, Keys.N}, new LinkSwordAttackCommand(game) },
                 {new Keys[] {Keys.G, Keys.C, Keys.V, Keys.B}, new MoveCameraCommand(game) },
                 {new Keys[] {Keys.E, Keys.P}, new PauseCommand(game)},
             };
